# services/api/openapi.json
{
  "components": {
    "schemas": {
      "AIActionProposalRead": {
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action_type",
          "payload",
          "explanation",
          "status",
          "created_at",
          "expires_at",
          "resolved_at"
        ],
        "title": "AIActionProposalRead",
        "type": "object"
      },
      "AIChatRead": {
        "properties": {
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIActionProposalRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          },
          "tool_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              }
            ],
            "title": "Tool Result"
          }
        },
        "required": [
          "response",
          "tool_name",
          "tool_result",
          "disclaimer"
        ],
        "title": "AIChatRead",
        "type": "object"
      },
      "AIChatRequest": {
        "properties": {
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AIChatRequest",
        "type": "object"
      },
      "AIProposalResolutionRead": {
        "properties": {
          "created_resource": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Resource"
          },
          "proposal": {
            "$ref": "#/components/schemas/AIActionProposalRead"
          }
        },
        "required": [
          "proposal"
        ],
        "title": "AIProposalResolutionRead",
        "type": "object"
      },
      "AccountCreate": {
        "properties": {
          "account_type": {
            "enum": [
              "cash",
              "checking",
              "savings",
              "credit_card",
              "prepaid",
              "investment",
              "loan",
              "wallet",
              "custom"
            ],
            "title": "Account Type",
            "type": "string"
          },
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "available_balance_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Balance Minor"
          },
          "balance_minor": {
            "default": 0,
            "title": "Balance Minor",
            "type": "integer"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "include_in_net_worth": {
            "default": true,
            "title": "Include In Net Worth",
            "type": "boolean"
          },
          "include_in_safe_to_spend": {
            "default": true,
            "title": "Include In Safe To Spend",
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "name",
          "account_type"
        ],
        "title": "AccountCreate",
        "type": "object"
      },
      "AccountRead": {
        "properties": {
          "account_type": {
            "title": "Account Type",
            "type": "string"
          },
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "available_balance_minor": {
            "title": "Available Balance Minor",
            "type": "integer"
          },
          "balance_minor": {
            "title": "Balance Minor",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "include_in_net_worth": {
            "title": "Include In Net Worth",
            "type": "boolean"
          },
          "include_in_safe_to_spend": {
            "title": "Include In Safe To Spend",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "account_type",
          "currency",
          "balance_minor",
          "available_balance_minor",
          "include_in_net_worth",
          "include_in_safe_to_spend",
          "archived",
          "notes",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "AccountRead",
        "type": "object"
      },
      "AccountUpdate": {
        "properties": {
          "account_type": {
            "anyOf": [
              {
                "enum": [
                  "cash",
                  "checking",
                  "savings",
                  "credit_card",
                  "prepaid",
                  "investment",
                  "loan",
                  "wallet",
                  "custom"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Type"
          },
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "include_in_net_worth": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include In Net Worth"
          },
          "include_in_safe_to_spend": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include In Safe To Spend"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "AccountUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Offline-sync-ready personal finance API. All monetary values are integer minor units.",
    "title": "MoneyPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/accounts": {
      "get": {
        "operationId": "list_accounts_api_v1_accounts_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccountRead"
                  },
                  "title": "Response List Accounts Api V1 Accounts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Accounts",
        "tags": [
          "accounts"
        ]
      },
      "post": {
        "operationId": "create_account_api_v1_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/accounts/{account_id}": {
      "delete": {
        "operationId": "delete_account_api_v1_accounts__account_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "accounts"
        ]
      },
      "get": {
        "operationId": "get_account_api_v1_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Account",
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "operationId": "update_account_api_v1_accounts__account_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Account",
        "tags": [
          "accounts"
        ]
      }
    },
    "/api/v1/ai/chat": {
      "post": {
        "operationId": "chat_api_v1_ai_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIChatRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/proposals": {
      "get": {
        "operationId": "list_proposals_api_v1_ai_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AIActionProposalRead"
                  },
                  "title": "Response List Proposals Api V1 Ai Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Proposals",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/proposals/{proposal_id}/approve": {
      "post": {
        "operationId": "approve_proposal_api_v1_ai_proposals__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProposalResolutionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Proposal",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/v1/ai/proposals/{proposal_id}/reject": {
      "post": {
        "operationId": "reject_proposal_api_v1_ai_proposals__proposal_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIProposalResolutionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Proposal",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}